#54 nonconformant themes mapping fixes
