+++ examples/themes_mapping.json
@@ -1,0 +1,1 @@
+{"meta":{"authority_file_src":"http:\/\/publications.europa.eu\/mdr\/authority\/data-theme\/","property":"dcat:theme","vocabulary":"topics"},"data":[{"name":"Agricoltura, pesca, silvicoltura e prodotti alimentari","syn":["agri","agricoltura","agricoltura-e-allevamento","agricoltura-e-pesca","agricoltura-e-sviluppo-rurale","agricoltura-pesca-politiche-forestali-e-alimentari","animali","attivit\u00e0 agricola","bevande e zucchero","coltivazione di terreni agricoli","foresta","industria agro-alimentare","mezzo di produzione agricola","orientamento produttivo agricolo","pesca","politica agricola","prodotto agricolo trasformato","prodotto alimentare","prodotto animale","prodotto vegetale","produzione e strutture agricole","settore-primario","tecnologia alimentare","Agricoltura, pesca, silvicoltura e prodotti alimentari"]},{"name":"Economia e finanze","syn":["econ","analisi economica","assicurazioni","attivita-produttive","bilanci","bilancio","chimica","commercializzazione","commercio internazionale","commercio-e-consumatori","concorrenza","consumo","contabilit\u00e0 nazionale","crescita economica","distribuzione commerciale","economia","economia monetaria","economia-finanze","economia-finanze-e-tributi","edilizia e lavori pubblici","elettronica ed elettrotecnica","finanza-e-controlli","finanze pubbliche e politica di bilancio","fiscalit\u00f2","forma giuridica di societ\u00e0","gestione amministrativa","gestione contabile-","imprese","industria del cuoio e dei prodotti tessili","industria del legno","industria meccanica","industrie varie","investimenti e finanziamenti","istituti finanziari e di credito","lavoro","lavoro-e-formazione","libera circolazione dei capitali","metallurgia e siderurgia","organizzazione aziendale","politica commerciale","politica economica","politica tariffaria","politiche e strutture industriali","prezzi","regione e politica regionale","relazioni monetarie","sardegna-turismo","scambi economici","struttura economica","sviluppo-economico-formazione-e-lavoro","terzo-settore","tipo d'impresa","turismo","Economia e finanze"]},{"name":"Istruzione, cultura e sport","syn":["educ","comunicazione","conoscenza","cultura","cultura e religione","culturaeturismo","documentazione","educazione-e-istruzione","informatica e trattamento dei dati","informazione ed elaborazione dell'informazione","insegnamento","istruzione","istruzione-cultura-e-sport","istruzione-e-lavoro","istruzione-e-sport","istruzione-formazione-e-diritto-allo-studio","istruzione1","organizzazione scolastica","ravenna-2019","ricerca-e-innovazione","sardegna-digital-library","scuola","sport","turismo-e-sport","turismo-e-tempo-libero","turismo-sport-e-tempo-libero","vita sociale - tempo libero","Istruzione, cultura e sport"]},{"name":"Energia","syn":["ener","energia","politica energetica","industria carboniera e mineraria","industria petrolifera","industria nucleare ed elettrica","energia dolce","Energia"]},{"name":"Ambiente","syn":["envi","ambiente","ambiente naturale","ambiente-e-meteo","ambiente-e-territorio","ambientenaturale","cartografia","cat-meteo","clima","dati-ambientali","dati-geografici","degrado ambientale","difesa-suolo","foreste-e-parchi","geodatabase","geoportale-sardegna","gestione-del-territorio","meteo","pianificazione-territoriale-comunale","politica dell'ambiente","superfici-territoriali","territorio","territorio-ambiente","Ambiente"]},{"name":"Governo e settore pubblico","syn":["gove","agenda-digitale-dell-umbria","amministrazione","amministrazione-comunale","amministrazione-trasparente","conti-pubblici-territoriali","costruzione europea","diritto dell'Unione Europea","fesr2007-2013","finanze dell'Unione europea","government-e-settore-pubblico","governo-e-settore-pubblico","istituzioni dell'Unione europea e funzione pubblica europea","lavori parlamentari","opere-pubbliche","Parlamento","partito politico","por-fesr-2014-2020","potere esecutivo e amministrazione pubblica","procedura elettorale e voto","pubblica-amministrazione","quadro politico","Trasparenza","urbanistica","vita politica e sicurezza pubblica","Governo e settore pubblico"]},{"name":"Salute","syn":["heal","promozione-della-salute","salute","salute-e-sanita","salute-e-sicurezza","sanita","sanita-e-servizi-sanitari","Salute"]},{"name":"Tematiche internazionali","syn":["intr","difesa","equilibrio internazionale","Nazioni Unite","organizzazioni europee","organizzazioni extraeuropee","organizzazioni mondiali","organizzazioni non governative","politica di cooperazione","politica internazionale","Tematiche internazionali"]},{"name":"Giustizia, sistema giuridico e sicurezza pubblica","syn":["just","diritti e libert\u00e0","diritto civile","diritto internazionale","diritto penale","emergenza-e-sicurezza","fonti e branche del diritto","giustizia","giustizia-e-sicurezza","organizzazione della giustizia","Sicurezza","Giustizia, sistema giuridico e sicurezza pubblica"]},{"name":"Regioni e citt\u00e0","syn":["REGI","Africa","America","Asia-Oceania","Europa","geografia economica","geografia politica","paesi e territori d'oltremare","regioni degli Stati membri dell'Unione europea","Regioni e citt\u00e0","regioni-e-citta","Regioni e citt\u00e0"]},{"name":"Popolazione e societ\u00e0","syn":["soci","amministrazione e remunerazione del personale","condizioni e organizzazioni del lavoro","dati-demografici","dati-statistici","dati-urbanistici","demografia","demografia e popolazione","elettorale","elezioni","elezioni-e-attivita-politica","famiglia","incidenti-stradali","lavoro-e-politiche-sociali","mercato del lavoro","migrazione","occupazione","politica","popolazione-e-societa","popolazione-e-sviluppo-demografico","protezione sociale","protezione-civile","puglia-sociale","quadro sociale","rapporti di lavoro e diritto del lavoro","sociale","societa","statistica","urbanistica ed edilizia","vita sociale","welfare","Popolazione e societ\u00e0"]},{"name":"Scienza e tecnologia","syn":["Tech","produzione","ricerca e propriet\u00e0 intellettuale","scienza-e-tecnologia","scienze naturali e applicate","scienze umane","tecnologia e regolamentazione tecnica","Scienza e tecnologia"]},{"name":"Trasporti","syn":["tran","infrastrutture-e-trasporti","mobilita","mobilita-e-sicurezza","mobilita-e-trasporti","organizzazione dei trasporti","politica dei trasporti","trasporti","trasporti aerei e spaziali","trasporti marittimi e fluviali","trasporti terrestri","viabilita","Trasporti"]}]}
